% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post.R
\name{post}
\alias{post}
\title{Generic function that posts a packet of information to the DataCamp pusher.}
\usage{
post(packet)
}
\description{
Generic function that posts a packet of information to the DataCamp pusher.
}

