% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post.R
\name{post_exercise.datacamp}
\alias{post_exercise.datacamp}
\title{Post exercise information to DataCamp}
\usage{
post_exercise.datacamp(e, current.row)
}
\arguments{
\item{e}{swirl environment}

\item{current.row}{The current.row in the swirl lesson}
}
\description{
Post exercise information to DataCamp
}

