% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post.R
\name{post_finished.datacamp}
\alias{post_finished.datacamp}
\title{Indicate to DataCamp that a lesson is finished}
\usage{
post_finished.datacamp(e)
}
\arguments{
\item{e}{swirl environment (only used for determining the implementation)}
}
\description{
Indicate to DataCamp that a lesson is finished
}

