% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_init.datacamp}
\alias{post_init.datacamp}
\title{Post information after initialization}
\usage{
post_init.datacamp(e)
}
\arguments{
\item{e}{swirl environment}
}
\description{
Post information after initialization
}

