% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post.R
\name{post_mult_question.datacamp}
\alias{post_mult_question.datacamp}
\title{Post MCQ to DataCamp - choices were already given in post_exercise}
\usage{
post_mult_question.datacamp(e, choices)
}
\arguments{
\item{e}{swirl environment}

\item{choices}{the options in the Multiple Choice Question}
}
\description{
Post MCQ to DataCamp - choices were already given in post_exercise
}

