% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post.R
\name{post_progress.datacamp}
\alias{post_progress.datacamp}
\title{Post progress information to DataCamp}
\usage{
post_progress.datacamp(e)
}
\arguments{
\item{e}{swirl environment}
}
\description{
Post progress information to DataCamp
}

