% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post.R
\name{post_result.datacamp}
\alias{post_result.datacamp}
\title{Post the result of a testable exercise to DataCamp}
\usage{
post_result.datacamp(e, passed, feedback, hint)
}
\arguments{
\item{e}{swirl environment}

\item{passed}{logical value indicating whether or not the corresponding exercise passed}

\item{feedback}{feedback message for student's answer on corresponding exercise}

\item{hint}{the hint swirl provides, can be NULL}
}
\description{
Post the result of a testable exercise to DataCamp
}

