% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swirlcamp.R
\name{swirl}
\alias{swirl}
\title{Start the DataCamp version of Swirl}
\usage{
swirl(init_filename = ".init.yaml", install = TRUE)
}
\arguments{
\item{init_filename}{yaml file containing initialization information. If not specified, defulats to .init.yaml}

\item{install}{logical value, whether or not to install course before starting swirl. Defaults to TRUE}
}
\description{
Start the DataCamp version of Swirl
}

