\name{swirlcamp-package}
\alias{swirlcamp-package}
\alias{swirlcamp}
\docType{package}
\title{
Learn R, in R, on DataCamp
}
\description{
A set of functions supporting swirl to use swirl on datacamp.
}
\details{
\tabular{ll}{
Package: \tab swirlcamp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-09\cr
License: \tab In-house use only for the moment\cr
}
# specify .yaml with all necessary info
initialize_swirl()
swirl()
}
\author{
Filip Schouwenaars <filip@datacamp.com>
Maintainer: Filip Schouwenaars <filip@datacamp.com>
}
\references{
Swirl <http://www.swirlstats.net>
}