% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swirlcamp.R
\name{swirl}
\alias{swirl}
\title{Start the DataCamp version of Swirl}
\usage{
swirl(init_filename = ".init.yaml", install = TRUE)
}
\arguments{
\item{install}{logical value, whether or not to install course before starting swirl. Defaults to TRUE}

\item{init_yaml}{yaml file containing initialization information. If not specified, defulats to .init.yaml}
}
\description{
TESTING COURSES: cp -R ~/courses/swirl_courses/R_Programming ~/.datacamp/Courses/
}
\details{
UPLOADING ZIP TO S3:
s3cmd put --acl-public ~/courses/swirl_courses/R_Programming.zip s3://assets.datacamp.com/course/swirl/
}

